+++ src/main/java/kuit/modi/dto/ai/request/AutoGenerateRequest.java
@@ -1,12 +1,14 @@
-package kuit.modi.dto.ai;
+package kuit.modi.dto.ai.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class LanguageStyleRequest {
-    private String summary;
+public class AutoGenerateRequest {
+    private List<String> tags;
 }
