+++ src/main/java/kuit/modi/dto/ai/request/LanguageStyleRequest.java
@@ -1,14 +1,12 @@
-package kuit.modi.dto.ai;
+package kuit.modi.dto.ai.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class AutoGenerateRequest {
-    private List<String> tags;
+public class LanguageStyleRequest {
+    private String summary;
 }
