+++ src/main/java/kuit/modi/dto/ai/request/SummaryRequest.java
@@ -1,4 +1,4 @@
-package kuit.modi.dto.ai;
+package kuit.modi.dto.ai.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
